finished all non admin things

+++ frontend/src/components/Delete/index.js
@@ -1,22 +1,54 @@
 // /frontend/src/components/Delete/index.js
 
 import { useDispatch, useSelector } from "react-redux";
+import * as approvalActions from "../../store/approval";
 import * as bookActions from "../../store/book";
+import * as userActions from "../../store/user";
 import { useHistory } from "react-router-dom";
 import { useEffect, useState } from "react";
+import { isValidUrl } from "../BookForm";
 
 
 const Delete = ({ params: { itemName, id }, setIsOpen }) => {
     const dispatch = useDispatch();
     const history = useHistory();
-    const err = useSelector(bookActions.selErr);
-    const msg = useSelector(bookActions.selMsg);
+    const bookErr = useSelector(bookActions.selErr);
+    const appErr = useSelector(approvalActions.selErr);
+    const bookMsg = useSelector(bookActions.selMsg);
+    const appMsg = useSelector(approvalActions.selMsg);
+    const user = useSelector(userActions.selUser);
+    const book = useSelector(bookActions.selBook);
+
+    const [reason, setReason] = useState('');
+    const [errors, setErrors] = useState([]);
 
     const handleYes = (e) => {
         e.preventDefault();
 
+        let mutBook = book ? {
+            bookId: book.id,
+            title: book.title,
+            subtitle: book.subtitle,
+            authors: book.authors,
+            publishYear: book.publishYear,
+            pageCount: book.pageCount,
+            synopsis: book.synopsis,
+            thumbnail: book.thumbnail,
+            reason: 'DELETE' + reason
+        } : {};
+
+        if (book.pdfLink && isValidUrl(book.pdfLink)) {
+            mutBook.pdfLink = book.pdfLink;
+        }
+
+
         if (itemName === 'book') {
-            dispatch(bookActions.deleteBook(id));
+            if (user && user.admin) {
+                dispatch(bookActions.deleteBook(id));
+            }
+            else {
+                dispatch(approvalActions.requestApproval(mutBook));
+            }
         }
 
     };
@@ -26,27 +58,52 @@ const Delete = ({ params: { itemName, id }, setIsOpen }) => {
         setIsOpen(false);
     };
 
-    useEffect(() => {
-        if (err && err.length) {
-            setIsOpen(false);
-            alert(err);
-        }
-    }, [err]);
+    const handleChange = (e) => {
+        e.preventDefault();
+        setReason(e.target.value);
+    }
 
     useEffect(() => {
-        if (msg && msg.length) {
-            history.push(`/${itemName}s`);
+        if (bookErr && bookErr.length) {
+            setErrors(bookErr);
         }
-    })
+
+        if (appErr && appErr.length) {
+            setErrors(appErr);
+        }
+    }, [bookErr, appErr, setErrors]);
+
+    useEffect(() => {
+        if (bookMsg && bookMsg.length) {
+            setIsOpen(false);
+            alert(bookMsg);
+            dispatch(bookActions.clearMsg());
+        }
+
+        if (appMsg && appMsg.length) {
+            setIsOpen(false);
+            alert(appMsg);
+            dispatch(approvalActions.clearMsg());
+        }
+    }, [bookMsg, appMsg, history]);
 
 
     return (
         <div className='delete-container'>
             <h1>Confirm Delete</h1>
-            <p>Are you sure you want to remove this {itemName}?</p>
-            <button className='yes-delete' onClick={handleYes}>Yes (Delete {itemName})</button>
-            <button className='no-delete' onClick={handleNo}>No (Keep {itemName})</button>
+            {
+                book && user &&
+                (<>
+                    <p>Are you sure you want to remove this {itemName}?</p>
+                    <ul>
+                        {errors.map(err => <li key={err}>{err}</li>)}
+                    </ul>
+                    <input type="text" className="reason" value={reason} onChange={handleChange} />
+                    <button className='yes-delete' onClick={handleYes}>Yes (Delete {itemName})</button>
+                    <button className='no-delete' onClick={handleNo}>No (Keep {itemName})</button>
 
+                </>)
+            }
         </div>
     );
 };
